WEB-841: Content for Licensing & Partners section
