+++ packages/numenta.com/pages/machine-intelligence-technology/licensing-and-partners/_metatags.md
@@ -1,5 +1,5 @@
 ---
-title: "Business Strategy & IP"
+title: "Licensing & Partners"
 keywords: Numenta patents
 description: We've built a patent portfolio that we believe covers some of the foundational principles of machine intelligence. We make our IP freely available for scientific use and non-commercial purposes.  Interested in a commercial deployment?  We offer an IP license that enables commercial use.
 image: ./images/business.png
