+++ packages/numenta.com/pages/machine-intelligence-technology/licensing-and-partners/images/grokstream.png
