--- packages/numenta.com/pages/partners/_metatags.md
@@ -1,5 +1,0 @@
----
-title: Partners
-keywords: Numenta Partners
-description: Numenta partners with organizations that are able to add an application layer to our neuroscience-based technology. We are flexible in structuring our partnerships in a way that's mutually beneficial. 
----
