--- packages/numenta.com/pages/partners/index.css
@@ -1,26 +1,0 @@
-/**
- * Numenta Web Platform and Sites source code
- * MIT License (see LICENSE.txt)
- * Copyright © 2005—2017 Numenta <http://numenta.com>
- */
-
-.columns {
-  composes: layout-columns from '../../static/assets/css/theme.css';
-}
-
-.content {
-  composes: layout-content from '../../static/assets/css/theme.css';
-}
-
-.aside {
-  composes: mt4 from 'tachyons';
-  composes: ph3 from 'tachyons';
-  composes: tc from 'tachyons';
-  composes: layout-aside from '../../static/assets/css/theme.css';
-}
-
-.logomark {
-  composes: center from 'tachyons';
-  composes: h4 from 'tachyons';
-  composes: w4 from 'tachyons';
-}
